feat: 진행중인 상담인지 체크하는 로직을 webSocket 연결에 추가
