+++ boomerang/src/main/java/boomerang/UiController.java
@@ -1,0 +1,23 @@
+package boomerang;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestMapping;
+
+@Controller
+@RequestMapping("/ui")
+public class UiController {
+
+    @GetMapping("/rooms/page")
+    public String getChatRoomsPage(Model model) {
+        return "chat_rooms";
+    }
+
+    @GetMapping("/room/{roomId}")
+    public String getChatRoomPage(@PathVariable Long roomId, Model model) {
+        model.addAttribute("roomId", roomId);
+        return "chat_room";
+    }
+}
