+++ boomerang/src/main/java/boomerang/global/config/WebSocketConfig.java
@@ -1,6 +1,7 @@
 package boomerang.global.config;
 
 import boomerang.chat.service.ChatRoomService;
+import boomerang.consultation.service.ConsultationService;
 import boomerang.global.handler.ChatWebSocketHandler;
 import boomerang.global.handler.WebSocketHandshakeInterceptor;
 import boomerang.global.utils.JwtUtil;
@@ -17,16 +18,18 @@ public class WebSocketConfig implements WebSocketConfigurer {
     private final JwtUtil jwtUtil;
     private final MemberService memberService;
     private final ChatRoomService chatRoomService;
+    private final ConsultationService consultationService;
 
-    public WebSocketConfig(JwtUtil jwtUtil, MemberService memberService, ChatRoomService chatRoomService) {
+    public WebSocketConfig(JwtUtil jwtUtil, MemberService memberService, ChatRoomService chatRoomService, ConsultationService consultationService) {
         this.jwtUtil = jwtUtil;
         this.memberService = memberService;
         this.chatRoomService = chatRoomService;
+        this.consultationService = consultationService;
     }
 
     @Override
     public void registerWebSocketHandlers(WebSocketHandlerRegistry registry) {
         registry.addHandler(new ChatWebSocketHandler(chatRoomService), "/ws/chat/*")
-                .addInterceptors(new WebSocketHandshakeInterceptor(jwtUtil, memberService, chatRoomService));
+                .addInterceptors(new WebSocketHandshakeInterceptor(jwtUtil, memberService, chatRoomService, consultationService));
     }
 }
